:arrow_right: Move Tile and Obstacle out of subfolder
