+++ src/map/Obstacle.js
