+++ src/map/Obstacle.scss
